+++ dotnet/src/EfSupport/QuerableSupport/RepositoryQuery.cs
@@ -1,0 +1,23 @@
+﻿using System.Runtime.CompilerServices;
+
+namespace System.Data.Fuse {
+
+  ///// <summary>
+  ///// (from 'FUSE-fx.EfSupport')
+  ///// </summary>
+  //public class RepositoryQuery<TEntity, TKey> {
+
+  //  public RepositoryQuery(IRepository<TEntity, TKey> repo) {
+  //  }
+
+  //}
+
+  //public static class RepositoryQueryExtensions {
+
+  //  public static RepositoryQuery<TEntity, TKey> Query<TEntity,TKey>(this IRepository<TEntity,TKey> repo) {
+  //    return new RepositoryQuery<TEntity, TKey>(repo);
+  //  }
+
+  //}
+
+}