+++ dotnet/src/RepositoryContract/GenericListSearchParams.cs
@@ -1,6 +1,9 @@
 
 namespace System.Data.Fuse {
 
+  /// <summary>
+  /// (from 'FUSE-fx.RepositoryContract')
+  /// </summary>
   public class GenericListSearchParams {
 
     public string EntityName { get; set; }