+++ dotnet/src/RepositoryContract/ICommitable.cs
@@ -1,5 +1,9 @@
-﻿namespace System.Data.Fuse {
+﻿
+namespace System.Data.Fuse {
 
+  /// <summary>
+  /// (from 'FUSE-fx.RepositoryContract')
+  /// </summary>
   public interface ICommitable {
 
     void BeginTransaction();
@@ -7,4 +11,5 @@
     void RollbackTransaction(); 
 
   }
+
 }