+++ dotnet/src/RepositoryContract/ICommitableRepository.cs
@@ -1,7 +1,13 @@
+
 namespace System.Data.Fuse {
 
+  /// <summary>
+  /// (from 'FUSE-fx.RepositoryContract')
+  /// </summary>
   public interface ICommitableRepository<TEntity, TKey>
     : IRepository<TEntity, TKey>, ICommitable
-    where TEntity : class { }
+    where TEntity : class { 
+  
+  }
 
 }