+++ dotnet/src/RepositoryContract/[old]/(from UDAS)/KeyValuesCollection.cs
@@ -4,7 +4,6 @@
 //using System.Text;
 //using System.Threading.Tasks;
 
-////TODO: rename namespace of the final versions to "System.Data.Fuse"
 //namespace System.Data.UDAS.v1.Models {
 //  public class KeyValuesCollection {
 //    public List<object[]> CollectionOfKeyValues { get; set; }