+++ dotnet/src/RepositoryContract/[old]/(from UDAS)/ListSearchParams.cs
@@ -1,5 +1,5 @@
 ﻿
-////TODO: rename namespace of the final versions to "System.Data.Fuse"
+
 //namespace System.Data.UDAS.v1.Models {
 
 //  public class ListSearchParamsBase {