+++ dotnet/src/RepositoryContract/[old]/(from UDAS)/PaginatedResponse.cs
@@ -4,7 +4,7 @@
 //using System.Text;
 //using System.Threading.Tasks;
 
-////TODO: rename namespace of the final versions to "System.Data.Fuse"
+
 //namespace System.Data.UDAS.v1.Models {
 //  public class PaginatedResponse<T> {
 //    public IList<T> Page { get; set; }