+++ dotnet/src/RepositoryContract/[old]/(from keinIch)/DeletionParams.cs
@@ -1,8 +1,8 @@
-namespace System.Data.Fuse {
+//namespace System.Data.Fuse {
 
-  public class DeletionParams {
-    public string EntityName { get; set; }
-    public object[][] IdsToDelete { get; set; }
-  }
+//  public class DeletionParams {
+//    public string EntityName { get; set; }
+//    public object[][] IdsToDelete { get; set; }
+//  }
 
-}
+//}