+++ dotnet/src/RepositoryContract/[old]/(from keinIch)/EntityRef.cs
@@ -1,4 +1,4 @@
-//TODO: rename namespace of the final versions to "System.Data.Fuse"
+
 //namespace System.Data.UDAS.v1.Models {
 //  public class EntityRef {
 //    public object[] KeyValues { get; set; }