+++ dotnet/src/RepositoryContract/[old]/(from keinIch)/GenericAddOrUpdateParams.cs
@@ -1,8 +1,8 @@
-﻿using System.Collections.Generic;
+﻿//using System.Collections.Generic;
 
-namespace System.Data.Fuse {
-  public class GenericAddOrUpdateParams {
-    public string EntityName { get; set; }
-    public Dictionary<string, object> Entity { get; set; }
-  }
-}
+//namespace System.Data.Fuse {
+//  public class GenericAddOrUpdateParams {
+//    public string EntityName { get; set; }
+//    public Dictionary<string, object> Entity { get; set; }
+//  }
+//}