+++ vers/changelog.md
@@ -8,9 +8,9 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* adjusted some details for the ExpressionTree
 
-
+  
 
 ## v 0.2.0
 released **2024-03-11**, including: